use jenkins connection to query user data as well
